+++ js/main.js
@@ -1,6 +1,6 @@
 import {similarAds} from './data.js';
 import {renderingAds} from './rendering-ads.js';
-import './form-state';
+import {disabledPage, activatedPage} from './page-state';
 
 const mapCanvas = document.querySelector('#map-canvas');
 const adsListFragment = document.createDocumentFragment();
@@ -10,3 +10,6 @@ similarAds.forEach((card) => {
 });
 
 mapCanvas.append(adsListFragment.children[0]);
+
+disabledPage();
+activatedPage();
